app_news_extractor_yogonet_com/requirements.txt
selenium
pandas
google-cloud-bigquery
datetime
pandas-gbq